K174: Fixed bugs introduced in homomorphism laws.
