+++ kosmos-laws/src/main/kotlin/org/vorpal/kosmos/laws/algebra/ConjugationLaws.kt
