+++ kosmos-laws/src/main/kotlin/org/vorpal/kosmos/laws/homomorphism/RingHomomorphismLaws.kt
@@ -1,7 +1,6 @@
 package org.vorpal.kosmos.laws.homomorphism
 
 import io.kotest.property.Arb
-import org.vorpal.kosmos.algebra.morphisms.RingHomomorphism
 import org.vorpal.kosmos.algebra.structures.Ring
 import org.vorpal.kosmos.core.Eq
 import org.vorpal.kosmos.core.render.Printable
@@ -16,8 +15,8 @@ import org.vorpal.kosmos.laws.TestingLaw
  *
  *       hom(1_A) = 1_B
  */
-class UnitalRingHomomorphismLaws<A : Any, B : Any>(
-    private val hom: RingHomomorphism<A, B>,
+class RingHomomorphismLaws<A : Any, B : Any>(
+    private val hom: (A) -> B,
     private val domain: Ring<A>,
     private val codomain: Ring<B>,
     arb: Arb<A>,
@@ -36,7 +35,7 @@ class UnitalRingHomomorphismLaws<A : Any, B : Any>(
             preservesIdentityLaw(
                 domainIdentity = domain.mul.identity,
                 codomainIdentity = codomain.mul.identity,
-                hom = hom::invoke,
+                hom = hom,
                 eqB = eqB,
                 prA = prA,
                 prB = prB,
